fixed bug where stuff didnt get undrawn

+++ core/src/de/fhbielefeld/pmdungeon/vorgaben/tools/Point.java
@@ -10,10 +10,10 @@ package de.fhbielefeld.pmdungeon.vorgaben.tools;
  */
 public class Point {
 
-	public int x;
-	public int y;
+	public float x;
+	public float y;
 
-	public Point(int x, int y) {
+	public Point(float x, float y) {
 		this.x = x;
 		this.y = y;
 	}
